05/27/2008 3.0.16a
SV_Select.cpp completed! Simple model calculates elevation mask based on earth blockage (function of altitude). Also added in ability to perform a warm start. This requires
that the file current.alm and lastpvt.txt reside in the directory of execution. current.alm is a YUMA format almanac available from the site
http://www.navcen.uscg.gov/GPS/almanacs.htm. The warm start requires that the PC clock is close to UTC and the time zone is set correctly. The warm start
can be enabled with the "-w" flag to the executable. Note that if the almanac or pvt load fails then the receiver will fall back to a cold start mode,
which is the default behaviour.V3.0.16a